implemented rbac on submitted content editor suggestion

+++ pages/api/submitted_content_editor_suggestion/[id]/index.tsx
@@ -1,9 +1,6 @@
 import prisma from "@/prisma/client";
-import submittedContentEditorSuggestionAbility from "@/services/ability/submittedContentEditorSuggestionAbility";
 import getServerUser from "@/services/getServerUser";
 import logger from "@/services/logger";
-import { ForbiddenError } from "@casl/ability";
-import { accessibleBy } from "@casl/prisma";
 import { User } from "@prisma/client";
 import type { NextApiRequest, NextApiResponse } from "next";
 import * as Yup from "yup";
@@ -16,12 +13,10 @@ export default async function handler(
 
   try {
     user = await getServerUser(req, res);
-    
   } catch (error) {
     logger.error(error);
     return res.status(401).json({ error: { message: "Unauthorized" } });
   }
-  const ability = submittedContentEditorSuggestionAbility({ user });
 
   const getHandler = async () => {
     try {
@@ -37,7 +32,6 @@ export default async function handler(
         await prisma.submittedContentEditorSuggestion.findFirstOrThrow({
           where: {
             AND: [
-              accessibleBy(ability).SubmittedContentEditorSuggestion,
               {
                 id: {
                   equals: id,
@@ -64,12 +58,15 @@ export default async function handler(
 
       await validator.validate(req.query);
 
-      ForbiddenError.from(ability).throwUnlessCan(
-        "delete",
-        "SubmittedContentEditorSuggestion"
-      );
-
       const { id } = validator.cast(req.query);
+
+      if (!user.isAdmin) {
+        return res.status(403).json({
+          error: {
+            message: "You are not allowed to perform this action",
+          },
+        });
+      }
 
       const submittedContentEditorSuggestion =
         await prisma.submittedContentEditorSuggestion.delete({
